refactor(invoices): extract sample context builders and period overview logic
- invoices/template_context.py: move _Obj helper class and all three
  _build_sample_*_context() functions out of views.py. These are pure
  data-construction helpers with no request dependency.

- invoices/period_overview.py: extract the participant/metering-readiness
  query loop from InvoiceViewSet.period_overview into compute_period_overview().
  The view now delegates after input validation and ZEV permission check.

- invoices/views.py: remove extracted code, add imports for new modules,
  drop unused imports (datetime, timedelta, dt_timezone, MeteringPoint,
  MeteringPointAssignment, MeterReading, F, DecimalField,
  INVOICE_TRANSLATIONS, CONTRACT_TRANSLATIONS, ANNUAL_TRANSLATIONS).

views.py: 1543 → ~1197 lines (-346)
